planning: remove CIPV
* it dupliates with existing planning functions
* it creates code logic diff between different perception configurations
